Split libcore/task.rs into submodules
